Satoshi my web initial
Has several iterations already, handles the document DOM looking for
pricing information and replacing it with bitcoin centric values.

Realtime pricing data is loaded on each refresh, and the dollar context
is guessed based upon the content of the page and the host of the content

original pricing data is available in the hover context.